+++ tools/import_validation/util.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 """Utility functions for the import validation tool."""
 
+import os
 import pandas as pd
 import re
 
@@ -108,3 +109,52 @@ def filter_dataframe(df: pd.DataFrame,
         matching_indices.update(df[combined_condition].index)
 
     return df.loc[sorted(list(matching_indices))]
+
+
+def _is_relative_local(path_val: str) -> bool:
+    """Checks if a path is a relative, local file path.
+
+    This function identifies path strings that represent local relative files
+    (e.g., 'golden_data/un_wpp.csv') as opposed to absolute paths. It filters
+    out non-strings, empty strings, and absolute local paths.
+
+    Args:
+        path_val: The file path string to evaluate.
+
+    Returns:
+        True if the path represents a relative, local file path; False otherwise.
+    """
+    if not isinstance(path_val, str) or not path_val:
+        return False
+    return not os.path.isabs(path_val)
+
+
+def _find_base_dir(start_path: str, target_sub_path: str) -> str | None:
+    """Helper to find a base directory containing a target sub-path by walking up.
+
+    Starting from the absolute directory of `start_path`, this function recursively
+    checks if `target_sub_path` exists in the current folder. If not, it walks up the 
+    parent directory tree up to 8 levels. This is crucial for resolving paths relative 
+    to import-specific golden directories when tests/validation are run from
+    different working directories (such as the repository root in CI/CD).
+
+    Args:
+        start_path: The file or directory path to start the upward search from.
+        target_sub_path: The name of the subdirectory or file (e.g., 'golden_data')
+            to search for within the parent tree.
+
+    Returns:
+        The absolute path of the directory containing `target_sub_path` if found,
+        or None if the root was reached or the 8-level limit was exceeded.
+    """
+    if not start_path:
+        return None
+    curr = os.path.abspath(start_path)
+    for _ in range(8):  # limit to 8 levels up
+        if os.path.exists(os.path.join(curr, target_sub_path)):
+            return curr
+        parent = os.path.dirname(curr)
+        if parent == curr:
+            break
+        curr = parent
+    return None
