+++ util/file_util.py
@@ -619,7 +619,7 @@ def file_write_csv_dict(py_dict: dict,
                         filename: str,
                         columns: list = None,
                         key_column_name: str = 'key') -> list:
-    """Returns the filename after writing py_dict with a csv row per item.
+    """Returns the list of columns after writing py_dict with a csv row per item.
 
   Each dictionary items is written as a row in the CSV file.
 
@@ -675,9 +675,7 @@ def file_write_csv_dict(py_dict: dict,
                         columns.append(col)
     if len(columns) == 1:
         # Check if values are dicts. If they are, it's not a primitive value.
-        is_value_dict = any(
-            isinstance(value, dict) for value in py_dict.values())
-        if not is_value_dict:
+        if not any(isinstance(value, dict) for value in py_dict.values()):
             # Value is not a dict. Write it as a column name value.
             value_column_name = 'value'
             columns.append(value_column_name)
